add Annotation configuration for Spring and Hibernate

+++ src/main/java/org/buldakov/employeeControl/DAO/DAOInterface.java
@@ -3,7 +3,7 @@ package org.buldakov.employeeControl.DAO;
 import java.util.List;
 
 public interface DAOInterface<E>{
-    E findById(int id);
+    E findById(long id);
 
     void save(E e);
